backend/requirements.txt
Django==2.1.4
Pillow
unidecode
cloudinary
django-filter
graphene_django
django-cors-headers
django-taggit
psycopg2
django-graphql-jwt
graphene-file-upload
djoser
djangorestframework
djangorestframework-jwt
